rendererGLES: small optimization, use short instead of int for systems with small memory amount to spare, refs #1100

+++ src/rendererGLES/tr_local.h
@@ -948,7 +948,7 @@ typedef struct srfTriangles_s
 
 	// triangle definitions
 	int numIndexes;
-	int *indexes;
+	short *indexes;
 
 	int numVerts;
 	drawVert_t *verts;
